move manualUnsavedChangesPopup.svelte.svelte in popups folder
